fix(fno): preserve spectral scatter precision
Allocate the inverse-transform Fourier buffer at the spectral convolution dtype so mixed float32 inputs and float64 weights retain complex128 precision without JAX scatter narrowing.

Add an x64 mixed-dtype regression that treats FutureWarning as fatal, make the real-data training warning fatal, and enroll the owning test in the strict MyPy and NumPy-docstring cohort.

Benchmarks: 25 focused FNO and controller tests passed under fatal FutureWarnings; strict MyPy passed across 651 files; expansion, docstring, Ruff, pre-commit, and claim-coverage gates passed.

Known Failures: coverage instrumentation for this JAX module retains the documented coverage-tracer circular-import blocker, so changed-line execution is proven by warning-fatal direct tests. The global structural-docstring audit retains 11 pre-existing validation gaps. No full local suite or hosted CI was run.

Seat: 3202544

Authored by Anulum Fortis & Arcane Sapience (protoscience@anulum.li)

+++ src/scpn_fusion/core/fno_jax_training.py
@@ -75,7 +75,7 @@ def fno_layer(
     out_ft_low = jnp.einsum("oimj,mji->mjo", weights, x_ft_low)
 
     # Pad back
-    out_ft = jnp.zeros_like(x_ft)
+    out_ft = jnp.zeros_like(x_ft, dtype=out_ft_low.dtype)
     out_ft = out_ft.at[:modes, :modes, :].set(out_ft_low)
 
     x_out = jnp.fft.irfft2(out_ft, axes=(0, 1), s=x.shape[:2])
